requirements.txt
websocket-client>=1.6.0
requests>=2.31.0
sentence-transformers>=2.2.2
chromadb>=0.4.0
fastapi>=0.104.0
uvicorn>=0.24.0
jinja2>=3.1.2
pyyaml>=6.0.1
jieba>=0.42.1
